+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v3.0.3
 ### Fixed
 - screen reader support for calendar
+- take in props to override browser locale for calendar
 
 ## v3.0.2
 ### Fixed